+++ src/Sulu/Bundle/AudienceTargetingBundle/Gruntfile.js
@@ -12,7 +12,11 @@ module.exports = function (grunt) {
     });
 
     // load all grunt tasks
-    require('matchdep').filterDev('grunt-*').forEach(grunt.loadNpmTasks);
+    require('matchdep').filterDev('grunt-*').forEach(function(name) {
+        if ('grunt-cli' !== name) {
+                grunt.loadNpmTasks(name);
+        }
+    });
 
     grunt.initConfig({
         pkg: grunt.file.readJSON('package.json'),
@@ -30,6 +34,13 @@ module.exports = function (grunt) {
             scripts: {
                 files: ['Resources/public/**'],
                 tasks: ['publish']
+            }
+        },
+        copy: {
+            templates: {
+                files: [
+                    {expand: true, cwd: srcpath, src: ['**/*.html'], dest: destpath}
+                ]
             }
         },
         compass: {
@@ -72,6 +83,8 @@ module.exports = function (grunt) {
     grunt.registerTask('build', [
         'uglify',
         'compass:dev',
-        'cssmin'
+        'cssmin',
+        'copy:templates',
+        'replace:build'
     ]);
 };