+++ src/Sulu/Bundle/SnippetBundle/Gruntfile.js
@@ -12,7 +12,11 @@ module.exports = function (grunt) {
     });
 
     // load all grunt tasks
-    require('matchdep').filterDev('grunt-*').forEach(grunt.loadNpmTasks);
+    require('matchdep').filterDev('grunt-*').forEach(function(name) {
+        if ('grunt-cli' !== name) {
+            grunt.loadNpmTasks(name);
+        }
+    });
 
     grunt.initConfig({
         pkg: grunt.file.readJSON('package.json'),
@@ -45,6 +49,19 @@ module.exports = function (grunt) {
                 }
             }
         },
+        replace: {
+            build: {
+                options: {
+                    variables: {
+                        'sulusnippet/js': 'sulusnippet/dist'
+                    },
+                    prefix: ''
+                },
+                files: [
+                    {src: ['Resources/public/dist/main.js'], dest: 'Resources/public/dist/main.js'}
+                ]
+            }
+        },
         cssmin: {
             // TODO: options: { banner: '<%= meta.banner %>' },
             compress: {
@@ -65,7 +82,8 @@ module.exports = function (grunt) {
     ]);
 
     grunt.registerTask('build:js', [
-        'uglify'
+        'uglify',
+        'replace:build'
     ]);
 
     grunt.registerTask('build', [